refactor: render entire suggestions section with React
Use a single directive the suggestion panel.
