+++ app/editor/editor-suggestions.html
@@ -6,7 +6,4 @@
      resizer-bottom="#editor-suggestions">
 </div>
 
-<aside ng-if="show" class="Editor-suggestions Editor-panel u-bgHigh" id="editor-suggestions" ng-class="{'is-search-active': search.isVisible}">
-  <suggestion-header></suggestion-header>
-  <suggestions-body></suggestions-body>
-</aside>
+<suggestions-panel></suggestions-panel>
